fix error : Create offer (refacto) + WIP : context loses its state when changing page
