+++ src/components/CreateOffer/style.jsx
@@ -1,0 +1,27 @@
+import styled from 'styled-components';
+
+export const FormContainer = styled.form`
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+`;
+
+export const Input = styled.input`
+  margin: 4px;
+`;
+export const InputButton = styled.input`
+  display: flex;
+  margin: 5px;
+  cursor: pointer;
+`;
+export const Label = styled.label`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
+export const Container = styled.div`
+  border: solid;
+  border-width: thin;
+  padding: 25px;
+  margin: 0 500px;
+`;
